requirements.txt
mne>=0.24.0
pyedflib>=0.1.15
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
torch>=1.9.0
pytorch-lightning>=1.5.0
tensorboard>=2.6.0
PyYAML>=5.4.0
jupyter>=1.0.0
notebook>=6.4.0